internal: make `expand_speculative` a method on `MacroCallId`
For symmetry with `MacroCallId::macro_expand`
